+++ dev/com.ibm.ws.jsf.2.2_fat/test-applications/JSF22ActionListener.war/src/com/ibm/ws/jsf22/fat/actionlistener/bean/faces40/TestSimpleBean.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2015, 2019 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ */
+package com.ibm.ws.jsf22.fat.actionlistener.bean.faces40;
+
+import java.io.Serializable;
+
+import javax.faces.bean.ManagedBean;
+import javax.faces.bean.SessionScoped;
+
+@ManagedBean
+@SessionScoped
+public class TestSimpleBean implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    private String value1;
+    private int value2 = 1;
+
+    public String getValue1() {
+        return value1;
+    }
+
+    public void setValue1(String value1) {
+        this.value1 = value1;
+    }
+
+    public int getValue2() {
+        return value2;
+    }
+
+    public void setValue2(int value2) {
+        this.value2 = value2;
+    }
+}
