+++ dev/com.ibm.ws.jsf.2.2_fat/test-applications/JSF22ActionListener.war/src/com/ibm/ws/jsf22/fat/actionlistener/bean/jsf22/TestSimpleBean.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  */
-package com.ibm.ws.jsf22.fat.actionlistener.bean;
+package com.ibm.ws.jsf22.fat.actionlistener.bean.jsf22;
 
 import java.io.Serializable;
